+++ spatial_samplers/sampler_shepard_quintic.h
@@ -87,7 +87,7 @@ public:
     {
         OTYPE vsum = 0;
         REAL  wsum = 0;
-        for( INT i = 0 ; i < data_points.size() ; i++ ) {
+        for( size_t i = 0 ; i < data_points.size() ; i++ ) {
             auto dist2 = normsq( focus - data_points[i].first );
             if( dist2 < r * r ) {
                 REAL w = quintic_polynomial( sqrt( dist2 ) );
